requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate
Flask-Bcrypt==0.7.1
psycopg2
python-dotenv==0.19.1
SQLAlchemy==1.4.23
werkzeug==2.0.3
